JRE Contribution
This change adds support for contributing a JRE to an application.  The JRE
contribution will occur if openjdk-jre is present in the build plan.  If the
build metadata key is set, the contribution consists of extracting the JRE to
a cache layer and setting JAVA_HOME.  If the launch metadata key is set, the
contribution consists of extracting the JRE to a launch layer and writing a
.profile.d file to set JAVA_HOME.