BP_JAVA_VERSION
Previously, only other buildpacks could choose the version of Java that would
be contributed to an application.  This change adds a user configuration
option in the form of the $BP_JAVA_VERSION variable that allows the user to
configure the version that will be contributed.  This value can be either an
explicit version that matches a version available in the buildpack, or more
likely will be a wildcard version such as 1.8.*.

[resolves #5]

Signed-off-by: Ben Hale <bhale@pivotal.io>