Substitute JDK for JRE
Currently, if no JRE can be found an error is returned.  Unfortunately there
are a number of OpenJDK distributions (Corretto, GraalVM) that do not release
JREs as a separate artifact.  This change adds the ability to substitute an
equivalent JDK in for the missing JRE, while displaying a large warning
message to the user about the security implications of this behavior.

Signed-off-by: Ben Hale <bhale@pivotal.io>