Sane Memory Limit
Previously, when the memory calculator attempted to detect the memory limit of
the container, the only check it would do was whether the container was
configured to be larger than 64TB to ensure that the JVM would get a number
that it would tolerate.  It didn't really handle the default unconfigured case
(8 Exabytes) in a way that most users expected, resulting in really slow
startup times.

This change updates the memory calculator script to notice that the container
memory limit is unconfigured and if it is, configures the memory calculator to
act like the container was limited to 1GB.

[resolves #11]

Signed-off-by: Ben Hale <bhale@pivotal.io>